<==== Mercurial data ====>
Branch: x_CLOUD-213_move_only_func

+++ common_libs/nx_media/src/nx/media/jpeg.cpp
@@ -44,12 +44,12 @@ QImage decompressJpegImage(const char *data, size_t size)
 
     return QImage(buffer, width, height, QImage::Format_ARGB32, &imageCleanup, buffer);
 }
-#else
+#else // USE_LIBJPEG
 QImage decompressJpegImage(const char *data, size_t size)
 {
-    return QImage::fromData(reinterpret_cast<const uchar*>(data), size);
+    return QImage::fromData(reinterpret_cast<const uchar*>(data), (int) size);
 }
-#endif
+#endif // USE_LIBJPEG
 
 QImage decompressJpegImage(const QByteArray &data)
 {
